requirements.txt
notion-client>=2.2.1,<3.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
PyYAML>=6.0
python-dateutil>=2.8.2
